+++ src/demo_tezos_domains/models.py
@@ -1,3 +1,4 @@
+from tortoise import ForeignKeyFieldInstance
 from tortoise import Model
 from tortoise import fields
 
@@ -9,7 +10,7 @@ class TLD(Model):
 
 class Domain(Model):
     id = fields.CharField(max_length=255, pk=True)
-    tld = fields.ForeignKeyField('models.TLD', 'domains')
+    tld: ForeignKeyFieldInstance[TLD] = fields.ForeignKeyField('models.TLD', 'domains')
     expiry = fields.DatetimeField(null=True)
     owner = fields.CharField(max_length=36)
     token_id = fields.BigIntField(null=True)
@@ -17,5 +18,5 @@ class Domain(Model):
 
 class Record(Model):
     id = fields.CharField(max_length=255, pk=True)
-    domain = fields.ForeignKeyField('models.Domain', 'records')
+    domain: ForeignKeyFieldInstance[Domain] = fields.ForeignKeyField('models.Domain', 'records')
     address = fields.CharField(max_length=36, null=True)