+++ src/demo_tzcolors/models.py
@@ -1,5 +1,6 @@
 from enum import IntEnum
 
+from tortoise import ForeignKeyFieldInstance
 from tortoise import Model
 from tortoise import fields
 
@@ -22,7 +23,7 @@ class Token(Model):
     amount = fields.BigIntField()
     level = fields.BigIntField()
     timestamp = fields.DatetimeField()
-    holder = fields.ForeignKeyField('models.Address', 'tokens')
+    holder: ForeignKeyFieldInstance[Address] = fields.ForeignKeyField('models.Address', 'tokens')
 
     class Meta:
         table = 'tokens'
@@ -30,10 +31,10 @@ class Token(Model):
 
 class Auction(Model):
     id = fields.BigIntField(pk=True)
-    token = fields.ForeignKeyField('models.Token', 'auctions')
+    token: ForeignKeyFieldInstance[Token] = fields.ForeignKeyField('models.Token', 'auctions')
     bid_amount = fields.BigIntField()
-    bidder = fields.ForeignKeyField('models.Address', 'winning_auctions')
-    seller = fields.ForeignKeyField('models.Address', 'created_auctions')
+    bidder: ForeignKeyFieldInstance[Address] = fields.ForeignKeyField('models.Address', 'winning_auctions')
+    seller: ForeignKeyFieldInstance[Address] = fields.ForeignKeyField('models.Address', 'created_auctions')
     end_timestamp = fields.DatetimeField()
     status = fields.IntEnumField(AuctionStatus)
     level = fields.BigIntField()
@@ -45,8 +46,8 @@ class Auction(Model):
 
 class Bid(Model):
     id = fields.BigIntField(pk=True)
-    auction = fields.ForeignKeyField('models.Auction', 'bids')
+    auction: ForeignKeyFieldInstance[Auction] = fields.ForeignKeyField('models.Auction', 'bids')
     bid_amount = fields.BigIntField()
-    bidder = fields.ForeignKeyField('models.Address', 'bids')
+    bidder: ForeignKeyFieldInstance[Address] = fields.ForeignKeyField('models.Address', 'bids')
     level = fields.BigIntField()
     timestamp = fields.DatetimeField()