+++ src/dipdup/dipdup.py
@@ -17,7 +17,6 @@ from typing import Tuple
 
 from prometheus_client import start_http_server  # type: ignore
 from tortoise.exceptions import OperationalError
-from tortoise.transactions import get_connection
 
 from dipdup.codegen import DipDupCodeGenerator
 from dipdup.config import ContractConfig
@@ -61,6 +60,7 @@ from dipdup.scheduler import SchedulerManager
 from dipdup.utils import is_importable
 from dipdup.utils import slowdown
 from dipdup.utils.database import generate_schema
+from dipdup.utils.database import get_connection
 from dipdup.utils.database import get_schema_hash
 from dipdup.utils.database import prepare_models
 from dipdup.utils.database import tortoise_wrapper
@@ -430,7 +430,7 @@ class DipDup:
     async def _initialize_schema(self) -> None:
         self._logger.info('Initializing database schema')
         schema_name = self._config.schema_name
-        conn = get_connection(None)
+        conn = get_connection()
 
         # NOTE: Try to fetch existing schema
         try: