+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 * jobs: Fixed jobs with `daemon` schedule never start.
 * jobs: Fixed failed jobs not throwing exceptions into the main loop.
 
+### Other
+
+* database: Tortoise ORM updated to `0.19.1`.
+
 ## 5.1.1 - 2022-05-13
 
 ### Fixed